GenericQueue: extract process_queue()
