+++ sqs_workers/queue.py
@@ -1,11 +1,15 @@
+import logging
 from typing import TYPE_CHECKING, Union
 
 from sqs_workers import codecs
 from sqs_workers.processor_mgr import ProcessorManagerProxy
+from sqs_workers.shutdown_policies import NEVER_SHUTDOWN
 
 if TYPE_CHECKING:
-    from sqs_workers import MemoryEnv
-    from sqs_workers import SQSEnv
+    from sqs_workers import MemoryEnv, SQSEnv
+
+
+logger = logging.getLogger(__name__)
 
 
 class GenericQueue(ProcessorManagerProxy):
@@ -39,3 +43,41 @@ class GenericQueue(ProcessorManagerProxy):
             _deduplication_id,
             _group_id,
         )
+
+    def add_raw_job(
+        self,
+        job_name,
+        message_body,
+        job_context,
+        content_type,
+        delay_seconds,
+        deduplication_id,
+        group_id,
+    ):
+        raise NotImplementedError("Must be implemented in subclass")
+
+    def process_queue(self, shutdown_policy=NEVER_SHUTDOWN, wait_second=10):
+        """
+        Run worker to process one queue in the infinite loop
+        """
+        logger.debug(
+            "Start processing queue {}".format(self.name),
+            extra={
+                "queue_name": self.name,
+                "wait_seconds": wait_second,
+                "shutdown_policy": repr(shutdown_policy),
+            },
+        )
+        while True:
+            result = self.process_batch(wait_seconds=wait_second)
+            shutdown_policy.update_state(result)
+            if shutdown_policy.need_shutdown():
+                logger.debug(
+                    "Stop processing queue {}".format(self.name),
+                    extra={
+                        "queue_name": self.name,
+                        "wait_seconds": wait_second,
+                        "shutdown_policy": repr(shutdown_policy),
+                    },
+                )
+                break
